requirements.txt
# copy và chạy code dưới trên cmd, terminal
# pip install -r requirements.txt

nextcord
mysql.connector
pygetwindow
pillow
pytz
humanize
importlib
requests